+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Features
     * `javax.faces.render.FacesRenderer`
     * `javax.faces.validator.FacesValidator`
     * `javax.faces.view.facelets.FaceletsResourceResolver`
+* Added possibility to specify which modules should be skipped (configuration parameter __`modulesToSkip`__)
 
 ###[code changes](https://github.com/ImmobilienScout24/deadcode4j/compare/deadcode4j-maven-plugin-1.3...master)